move MetadataReportConstants back into metata-report-api module
